Upgrading the commons-* dependencies and the log4j dependency.
I've used the license file from within each foo.jar to create a foo.license file - technically that's the bit that is the license and there's nothing that says the license should be the same between each version; things get tacked on the end if there are other bits in there (see the HTTP Server license). 

'ant clean test' passes successfully, which is about all I can think of to confirm things aren't screwed up. The only major API change is with Commons Modeler, here are the release notes: http://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/tags/MODELER_2_0/RELEASE-NOTES.txt

I removed the lib/readme.txt as it was just there as a CVS placeholder.

This is described in QUARTZ-451